Refactor: some class renaming and docker-compose file adding with kafka
   And also enable event dispatching in CassandraStoreDataManager
