+++ src/main/java/org/commonjava/indy/service/repository/change/event/kafka/StoreEventDeserializer.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2011-2021 Red Hat, Inc. (https://github.com/Commonjava/indy)
+ * Copyright (C) 2020 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.commonjava.indy.service.repository.event.kafka;
+package org.commonjava.indy.service.repository.change.event.kafka;
 
 import io.quarkus.kafka.client.serialization.ObjectMapperDeserializer;
-import org.commonjava.indy.service.repository.model.StoreKey;
+import org.commonjava.event.store.IndyStoreEvent;
 
-public class StoreKeyMapperDeserializer extends ObjectMapperDeserializer<StoreKey>
+public class StoreEventDeserializer
+        extends ObjectMapperDeserializer<IndyStoreEvent>
 {
-    public StoreKeyMapperDeserializer( Class<StoreKey> type )
+    public StoreEventDeserializer()
     {
-        super( type );
+        super( IndyStoreEvent.class );
     }
 }
