+++ src/test/java/org/commonjava/indy/service/repository/ftests/admin/event/AbstractStoreEventTest.java
@@ -18,7 +18,7 @@ package org.commonjava.indy.service.repository.ftests.admin.event;
 import io.smallrye.reactive.messaging.connectors.InMemoryConnector;
 import io.smallrye.reactive.messaging.connectors.InMemorySink;
 import org.commonjava.event.store.IndyStoreEvent;
-import org.commonjava.indy.service.repository.change.event.KafkaEventUtils;
+import org.commonjava.indy.service.repository.change.event.kafka.KafkaEventUtils;
 import org.commonjava.indy.service.repository.ftests.AbstractStoreManagementTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
