+++ src/test/java/org/commonjava/indy/service/repository/ftests/admin/event/CreateThenDeleteRepoWithEventTest.java
@@ -21,11 +21,8 @@ import io.smallrye.reactive.messaging.connectors.InMemorySink;
 import org.commonjava.event.store.IndyStoreEvent;
 import org.commonjava.event.store.StoreEventType;
 import org.commonjava.event.store.StorePostDeleteEvent;
-import org.commonjava.event.store.StorePostUpdateEvent;
 import org.commonjava.event.store.StorePreDeleteEvent;
-import org.commonjava.event.store.StorePreUpdateEvent;
-import org.commonjava.event.store.StoreUpdateType;
-import org.commonjava.indy.service.repository.change.event.KafkaEventUtils;
+import org.commonjava.indy.service.repository.change.event.kafka.KafkaEventUtils;
 import org.commonjava.indy.service.repository.ftests.matchers.RepoEqualMatcher;
 import org.commonjava.indy.service.repository.ftests.profile.ISPNFunctionProfile;
 import org.commonjava.indy.service.repository.model.Group;
@@ -38,7 +35,6 @@ import java.util.List;
 import static io.restassured.RestAssured.delete;
 import static io.restassured.RestAssured.given;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
-import static javax.ws.rs.core.Response.Status.OK;
 import static org.awaitility.Awaitility.await;
 import static org.commonjava.indy.service.repository.model.pkg.MavenPackageTypeDescriptor.MAVEN_PKG_KEY;
 import static org.hamcrest.CoreMatchers.is;
